+++ helpers/helpers_xxx_UI_chars.js
@@ -1,0 +1,69 @@
+﻿'use strict';
+
+// Font Awesome
+// https://www.fontawesomecheatsheet.com/
+const chars = {
+	// Web
+	lastFm			: '\uf202',
+	link			: '\uf0c1',
+	// Standard
+	close 			: '\uf00D',
+	ratingOn 		: '\uf005',
+	ratingOff		: '\uf006',
+	heartOn			: '\uf004',
+	heartOff		: '\uf08A',
+	music			: '\uf001',
+	headphones 		: '\uf025',
+	microphone 		: '\uf130',
+	noMicrophone 	: '\uf131',
+	minus			: '\uf068',
+	plus			: '\uf067',
+	wand			: '\uf0d0',
+	wrench			: '\uf0ad',
+	calendar		: '\uf073',
+	// Files
+	folderOpenBlack	: '\uf07c',
+	folderOpenWhite	: '\uf115',
+	emptyFileWhite	: '\uf016',
+	emptyFilesWhite	: '\uf0c5',
+	fileWhite		: '\uf0f6',
+	emptyFileBlack	: '\uf15b',
+	fileBlack		: '\uf15c',
+	save 			: '\uf0c7',
+	// Utilities
+	search			: '\uf002',
+	menu			: '\uf0C9',
+	console			: '\uf120',
+	preferences 	: '\uf013',
+	searchPlus 		: '\uf00e',
+	searchMinus		: '\uf010',
+	tag				: '\uf02b',
+	tags 			: '\uf02c',
+	duplicates 		: '\uf0c5',
+	filter 			: '\uf0b0',
+	info 			: '\uf05A',
+	// Playback
+	prev			: '\uf049',
+	next			: '\uf050',
+	play			: '\uf04B',
+	pause			: '\uf04C',
+	stop			: '\uf04D',
+	// Carets
+	up 				: '\uf077',
+	down 			: '\uf078',
+	// Arrows
+	downLongArrow	: '\uf175',
+	upLongArrow		: '\uf176',
+	leftLongArrow	: '\uf177',
+	rightLongArrow	: '\uf178',
+	levelDown 		: '\uf149',
+	levelUp 		: '\uf148',
+	prevCircle 		: '\uf0e2',
+	// Sort
+	sortAlphaAsc	: '\uf15d',
+	sortAlphaDes	: '\uf15e',
+	sortBarsAsc		: '\uf160',
+	sortBarsDes		: '\uf161',
+	sortNumAsc		: '\uf162',
+	sortNumDes		: '\uf163',
+};